Fixed Footer Sitting on Top of Map
I made some changes so that the footer did not sit on top of the map.
This was a problem because the map bounds and map center were not
correct to what the user sees.  I noticed this error as I started to
work on the ability to save the extent of map.